refactor: calling BatchProgress only Progress from now on
