+++ src/modalities/logging_broker/messages.py
@@ -29,8 +29,8 @@ class ExperimentStatus(Enum):
 
 
 @dataclass
-class BatchProgressUpdate:
-    """Object holding the state of the current batch computation progress."""
+class ProgressUpdate:
+    """Object holding the state of the current batch / step computation progress."""
 
     num_steps_done: int
     # Note: in case of ExperimentState.TRAIN, dataset_batch_id=global_train_batch_id
